Makes corruption recover by a percent instead of all
Right now, the recover function appears to expect a decimal, but gets a whole number. And so it ends up multiplying the corruption level by a negative number, which then gets clamped to 0. Which means that every single instance of butterfingers or corrupt influence caused by corruption, will instantly wipe away all corruption (Alicorn Amulet butterfingers/corrupt influence is different)

This change makes it a float/decimal when calling `recover`, so that corruption is reduced by what I assume is the intended percentages (25% for butterfingers, 10% per mob for corrupt influence)